+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.3 (October 21, 2017) [by noo-zh](https://github.com/noo-zh/ezcrop)
+
+### Features
+
+* Dynamic `exportZoom` using `exportWidth`
+* Production bundle optimization
+
+## 0.0.2 (?)
+
+Not documented.
+
 ## 0.0.1 (December 26, 2016)
 
 ### Features